fix: Fix all lint warnings coming from the new Jest ESLint plugin
