+++ packages/devtools/test/omnigraph/types.test.ts
@@ -1,4 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
+/* eslint-disable jest/expect-expect */
 import { OmniEdge, OmniNode, OmniPoint, OmniVector } from '@/omnigraph'
 import { Factory } from '@/types'
 import { EndpointId } from '@layerzerolabs/lz-definitions'
@@ -133,21 +134,11 @@ describe('schema/types', () => {
             const factory: Factory<[boolean], boolean> = (a: string): string => a
         })
 
-        it('should not allow a factory that changes input types', () => {
-            // @ts-expect-error The input of the factory is a string, not a boolean
-            const factory: Factory<[boolean], boolean> = async (a: string): boolean => !!a
-        })
-
-        it('should not allow a factory that changes output types', () => {
-            // @ts-expect-error The output of the factory is a string, not a boolean
-            const factory: Factory<[boolean], boolean> = async (a: string): string => a
-        })
-
         it('should allow a sync factory', () => {
             const factory: Factory<[boolean], boolean> = (a: boolean) => a
         })
 
-        it('should allow an sync factory', () => {
+        it('should allow an async factory', () => {
             const factory: Factory<[boolean], boolean> = async (a: boolean) => a
         })
     })
